+++ src/main/java/com/telepathicgrunt/world_blender/WorldBlenderClient.java
@@ -3,18 +3,21 @@ package com.telepathicgrunt.world_blender;
 import com.telepathicgrunt.world_blender.blocks.WBBlocks;
 import com.telepathicgrunt.world_blender.blocks.WBPortalBlockEntity;
 import com.telepathicgrunt.world_blender.blocks.WBPortalBlockEntityRenderer;
+import com.telepathicgrunt.world_blender.dimension.WBSkyProperty;
+import com.telepathicgrunt.world_blender.mixin.SkyPropertiesAccessor;
 import com.telepathicgrunt.world_blender.utils.GoVote;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendereregistry.v1.BlockEntityRendererRegistry;
 import net.fabricmc.fabric.api.network.ClientSidePacketRegistry;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
 
 public class WorldBlenderClient implements ClientModInitializer {
 	@Override
 	public void onInitializeClient() {
 		GoVote.init();
-
+		SkyPropertiesAccessor.getBY_IDENTIFIER().put(new Identifier(WorldBlender.MODID, "sky_property"), new WBSkyProperty());
 		BlockEntityRendererRegistry.INSTANCE.register(WBBlocks.WORLD_BLENDER_PORTAL_BE, WBPortalBlockEntityRenderer::new);
 
 		// Set cooldown for portal after server says it was triggered