+++ src/main/java/com/telepathicgrunt/world_blender/the_blender/FeatureGrouping.java
@@ -214,6 +214,9 @@ public class FeatureGrouping
 						allFeatures.append(jsonConfigObject.get("type").toString()).append(" ");
 						return allFeatures.toString();
 					}
+					else{
+						return allFeatures.toString();
+					}
 				}
 			}
 			else if(jsonConfigObject.has("feature")){
@@ -233,7 +236,7 @@ public class FeatureGrouping
 			return jsonStartObject.get("type").toString();
 		}
 
-		return "";
+		return jsonStartObject.toString();
 	}