+++ CHANGELOG.md
@@ -6,6 +6,15 @@ Welcome to the Github! If you are looking for the most recent stable version, th
 ------------------------------------------------
 #    | World Blender changelog |
 
+## (V.1.0.8 Changes) (1.16.3 Minecraft)
+
+##### Dimension:
+- Fixed sky color.
+ 
+##### Features:
+- Fixed Coral not spawning in dimension.
+
+  
 ## (V.1.0.7 Changes) (1.16.3 Minecraft)
      
 ##### Portal: